pyproject.toml
[project]
name = "aws-cdk-lightsail"
version = "0.1.0"
description = "Proyecto CDK (Python + uv) para aprovisionar Lightsail con IP estática"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib==2.*",
    "constructs==10.*",
]
